renamed Convertfilecommand
